Introduction and adding README
This is me adding introduction to what all of these folders are, organization is very important to make the workflow much easier.

This all be can rewritten and fixed upon further discussion so keep in mind it is not final or an absolute word.

Keep in mind this repo will be submitted as the project so keep all commits appropriate